+++ crates/star-cloudburst/CHANGELOG.md
@@ -1,8 +1,11 @@
+# 0.7.0
+- Rename `benitor` to `star-torrent` and split out `benitor` poop.
+
 # 0.6.0
 - Simplify and improve bin tester by deduplicating code to print results and iterating over `Result<Vec<PathBuf, _>>` instead of `Result<Vec<Torrent, _>>>`.
 - Add pretty colors to output.
 - Use wrappers around hash types in `Info` and everywhere else.
-- Implement a depth first iterator for `FileTree`.
+- Implement a depth first iterator for `FileTree` and associated tests.
 
 # 0.5.0
 - Wrap `ByteBuf` into `HexBytes`